+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Adds the `MultiSelect`, `MultiSelectGroup`, `FilterBar` component.
 - Adds the `useMultiSelect`, `useFilterBar` hook.
 - Adds the `onChangeEnd` prop in the `Slider` component.
+- Adds `autocomplete` attribute to `TextInput`s with type `"tel"`, `"email"`,
+  and `"url"`.
 
 ### Updates
 
@@ -27,6 +29,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 - Fixes spacing and alignment issues in the `FeedbackBox` component.
 - Fixes the width of the Privacy Policy link in the `FeedbackBox` component.
+- Fixes issue where `aria-describedby` attribute was not being added to
+  some `TextInput`s.
 
 ### Removes